\name{Breast Epithelial Subtypes}
\alias{BreastEpithelialSubtypes}
\title{BreastEpithelialSubtypes}
\description{
Usage:\cr
estProps <- BreastEpithelialSubtypes(DNAmeData, RNAseqData, hierarchical=T)\cr
\cr
Where:\cr
'estProps' is the estimated proportions of the important cell-types of the breast.\cr
'DNAmeData' is the DNA methylation microarray data-set of interest (with features/CpGs in rows, and samples in columns).\cr
'RNAseqData' is the RNA-seq data-set (with features/genes in rows, and samples in columns).
At least one of 'DNAmeData' and 'RNAseqData' must be provided. If both 'DNAmeData' and 'RNAseqData' are provided, their columns must correspond to the same tissue samples.
'hierarchical' determines whether to also estimate the proportions of the breast epithelial subtypes in relation to fat, adipose, and immune cells. It is only available if DNAmeData is provided.
}
\details{
Estimates breast tissue cell-type proportions from DNA methylation microarray data, including the breast epithelium cell-subtypes luminal progenitor, mature luminal, and basal, as described in: https://www.biorxiv.org/content/10.1101/2021.05.17.444493
}
\author{
Thomas Bartlett.
}
\references{

Bartlett, T. E., Chandna, S., & Roy, S. (2021). Inference of tissue relative proportions of the breast epithelial cell types luminal progenitor, basal, and luminal mature. https://www.biorxiv.org/content/10.1101/2021.05.17.444493

Pellacani, D., Bilenky, M., Kannan, N., Heravi-Moussavi, A., Knapp, D. J., Gakkhar, S., ... & Marra, M. A. (2016). Analysis of normal human mammary epigenomes reveals cell-specific active enhancer states and associated transcription factor networks. Cell reports, 17(8), 2060-2074.

Teschendorff, A. E., Breeze, C. E., Zheng, S. C., & Beck, S. (2017). A comparison of reference-based algorithms for correcting cell-type heterogeneity in Epigenome-Wide Association Studies. BMC bioinformatics, 18(1), 105.

The results published here are in whole or part based upon data generated by The Canadian Epigenetics, Epigenomics, Environment and Health Research Consortium (CEEHRC) initiative funded by the Canadian Institutes of Health Research (CIHR), Genome BC, and Genome Quebec. Information about CEEHRC and the participating investigators and institutions can be found at http://www.cihr-irsc.gc.ca/e/43734.html
}
\keyword{}
\seealso{
}
\examples{
}
